Fix an order-dependent flaky test (#1747)
Flaky test: BootstrapEnvironmentTest#assertGetFrameworkHostPort
Expectd value is "localhost", but the actual value is "127.0.0.1"

[Problem]
When the BootstrapEnvironment singleton is initialized,
the "properties" object is populated from elasticjob-cloud-scheduler.properties,
and the "hostname" becomes "127.0.0.1", while the default "hostname"
is "localhost". At this moment, assertGetFrameworkHostPort should
fail. But in some other tests, the "properties" object resets. So
in normal order, this test doesn't fail.

We can easily fail this test by simply running this test alone.

[Solution]
Reset the "properties", make sure we get the default value.